Launcher3: Change resource type for blur radius in DepthController
This allows us to define the blur radius in device-independent pixels
instead of physical pixels, which is necessary to fix discrepancies in
apparent blur strength across devices with different pixel densities.

+++ quickstep/src/com/android/launcher3/statehandlers/DepthController.java
@@ -126,7 +126,7 @@ public class DepthController implements StateHandler<LauncherState>,
     /**
      * Blur radius when completely zoomed out, in pixels.
      */
-    private int mMaxBlurRadius;
+    private float mMaxBlurRadius;
     private boolean mCrossWindowBlursEnabled;
     private WallpaperManager mWallpaperManager;
     private SurfaceControl mSurface;
